requirements.txt
fastapi>=0.115
uvicorn>=0.30
pydantic>=2.7
httpx>=0.27
tenacity>=8.5
pybreaker>=1.0.2
click>=8.1
prometheus-client>=0.20
pyyaml>=6.0
pytest>=8.0
requests>=2.32
docker>=6.0
